using torch everywhere for full gradient-enabled lettuce

+++ lettuce/equilibrium.py
@@ -21,7 +21,7 @@ class QuadraticEquilibrium(Equilibrium):
         return NativeQuadraticEquilibrium()
 
     def __call__(self, rho, u, *args):
-        exu = torch.tensordot(self.lattice.e, u, dims=1)
+        exu = torch.tensordot(self.lattice.e.to(self.lattice.dtype), u, dims=1)
         uxu = self.lattice.einsum("d,d->", [u, u])
         feq = self.lattice.einsum(
             "q,q->q",
@@ -50,9 +50,9 @@ class QuadraticEquilibrium_LessMemory(Equilibrium):
         return self.lattice.einsum(
             "q,q->q",
             [self.lattice.w,
-             rho * ((2 * torch.tensordot(self.lattice.e, u, dims=1) - self.lattice.einsum("d,d->", [u, u]))
+             rho * ((2 * torch.tensordot(self.lattice.e.to(self.lattice.dtype), u, dims=1) - self.lattice.einsum("d,d->", [u, u]))
                     / (2 * self.lattice.cs ** 2)
-                    + 0.5 * (torch.tensordot(self.lattice.e, u, dims=1) / (self.lattice.cs ** 2)) ** 2 + 1)]
+                    + 0.5 * (torch.tensordot(self.lattice.e.to(self.lattice.dtype), u, dims=1) / (self.lattice.cs ** 2)) ** 2 + 1)]
         )
